+++ src/strongbox/core.clj
@@ -1,11 +1,12 @@
 (ns strongbox.core
   (:require
+   [clojure.java.io]
    [clojure.set :refer [rename-keys]]
    [clojure.string :refer [lower-case starts-with? ends-with? trim]]
    [taoensso.timbre :as timbre :refer [debug info warn error report spy]]
    [clojure.spec.alpha :as s]
    [orchestra.core :refer [defn-spec]]
-   [taoensso.tufte :as tufte :refer [p profile]]
+   [taoensso.tufte :as tufte :refer [p]]
    [me.raynes.fs :as fs]
    [trptcolin.versioneer.core :as versioneer]
    [envvar.core :refer [env]]
@@ -16,15 +17,46 @@
     [zip :as zip]
     [http :as http]
     [logging :as logging]
-    [utils :as utils :refer [join nav-map nav-map-fn delete-many-files! static-slurp expand-path if-let*]]
+    [utils :as utils :refer [join nav-map nav-map-fn delete-many-files! expand-path if-let*]]
     [catalogue :as catalogue]
     [specs :as sp]
-    [joblib :as joblib]]))
+    [joblib :as joblib]])
+  (:import
+   [org.apache.commons.compress.compressors CompressorStreamFactory CompressorException]))
 
 (def default-config-dir "~/.config/strongbox")
 (def default-data-dir "~/.local/share/strongbox")
 
 (def num-concurrent-downloads (-> (Runtime/getRuntime) .availableProcessors))
+
+(defn-spec compressed-slurp (s/or :ok bytes?, :no-resource nil?)
+  "returns the bz2 compressed bytes of the given resource file `resource`.
+  returns `nil` if the file can't be found."
+  [resource string?]
+  (let [input-file (clojure.java.io/resource resource)]
+    (when input-file
+      (with-open [out (java.io.ByteArrayOutputStream.)]
+        (with-open [cos (.createCompressorOutputStream (CompressorStreamFactory.) CompressorStreamFactory/BZIP2, out)]
+          (clojure.java.io/copy (clojure.java.io/input-stream input-file) cos))
+        ;; compressed output stream (cos) needs to be closed to flush any remaining bytes
+        (.toByteArray out)))))
+
+(defn-spec decompress-bytes (s/or :ok string?, :nil-or-empty-bytes nil?)
+  "decompresses the given `bz2-bytes` as bz2, returning a string.
+  if bytes are empty or `nil`, returns `nil`.
+  if bytes are not bz2 compressed, throws an `IOException`."
+  [bz2-bytes (s/nilable bytes?)]
+  (when-not (empty? bz2-bytes)
+    (with-open [is (clojure.java.io/input-stream bz2-bytes)]
+      (try
+        (with-open [cin (.createCompressorInputStream (CompressorStreamFactory.) CompressorStreamFactory/BZIP2, is)]
+          (slurp cin))
+        (catch CompressorException ce
+          (throw (.getCause ce)))))))
+
+(def static-catalogue
+  "a bz2 compressed copy of the full catalogue used when the remote catalogue is unavailable or corrupt."
+  (compressed-slurp "full-catalogue.json"))
 
 (defn generate-path-map
   "generates filesystem paths whose location may vary based on the current working directory and environment variables.
@@ -56,6 +88,7 @@
                   :data-dir data-dir
                   :catalogue-dir data-dir
 
+                  ;; 2021-09-16: no longer used, exists only for cleanup and will be removed in the future.
                   ;; /home/$you/.local/share/strongbox/profile-data
                   :profile-data-dir (join data-dir "profile-data")
 
@@ -270,7 +303,6 @@
        (swap! state update-in [:cfg :addon-dir-list] conj stub)))
    nil))
 
-;; see also: strongbox.ui.cli/set-addon-dir! 
 (defn-spec set-addon-dir! nil?
   "adds a new :addon-dir to :addon-dir-list (if it doesn't already exist) and marks it as selected"
   [addon-dir ::sp/addon-dir]
@@ -279,7 +311,7 @@
         default-game-track (if (clojure.string/index-of addon-dir "_classic_") :classic :retail)]
     (dosync ;; necessary? makes me feel better
      (add-addon-dir! addon-dir default-game-track default-game-track-strictness)
-     ;; todo: this is UI logic ... consider moving to ui.cli
+     (swap! state assoc :selected-addon-list [])
      (swap! state assoc-in [:cfg :selected-addon-dir] addon-dir)
      (swap! state assoc :tab-list []))) ;; todo: consider adding a watch for this as well
   nil)
@@ -305,8 +337,8 @@
   "returns the addon-dir map for the given `addon-dir`, if it exists in the map.
   when called without args, returns the addon-dir map for the currently selected addon-dir."
   ([]
-   (when-let [addon-dir (selected-addon-dir)]
-     (addon-dir-map addon-dir)))
+   (assert-started)
+   (addon-dir-map (selected-addon-dir)))
   ([addon-dir (s/nilable ::sp/addon-dir)]
    (let [addon-dir-list (get-state :cfg :addon-dir-list)]
      (when (and addon-dir
@@ -329,7 +361,8 @@
      nil)))
 
 (defn-spec get-game-track (s/or :ok :addon-dir/game-track, :missing nil?)
-  "returns the game track for the given `addon-dir` or the currently selected addon-dir if no `addon-dir` given"
+  "returns the game track for the given `addon-dir`.
+  uses the currently selected addon-dir if no `addon-dir` given."
   ([]
    (get-game-track (selected-addon-dir)))
   ([addon-dir (s/nilable ::sp/addon-dir)]
@@ -366,9 +399,8 @@
   [state-atm ::sp/atom]
   (when (debug-mode?)
     (if-not @state-atm
-      (warn "application has not been started, no location to write log or profile data")
+      (warn "application has not been started, no location to write log data")
       (do
-        (logging/add-profiling-handler! (paths :profile-data-dir))
         (logging/add-file-appender! (paths :log-file))
         (info "writing logs to:" (paths :log-file))))))
 
@@ -427,8 +459,6 @@
   (let [final-config (config/load-settings cli-opts (paths :cfg-file) (paths :etag-db-file))]
     (swap! state merge final-config)
     (reset-logging!)
-    (when (contains? cli-opts :profile?)
-      (swap! state assoc :profile? (:profile? cli-opts)))
     (when (contains? cli-opts :spec?)
       (utils/instrument (:spec? cli-opts))))
   nil)
@@ -481,8 +511,8 @@
   "downloads `addon` to the given `install-dir`.
   see `download-addon-guard` for a version with checks."
   [addon :addon/installable, install-dir ::sp/writeable-dir]
-  (info (format "downloading '%s' version '%s'" (:label addon) (:version addon)))
   (when (expanded? addon)
+    (info (format "downloading '%s' version '%s'" (:label addon) (:version addon)))
     (let [output-fname (addon/downloaded-addon-fname (:name addon) (:version addon)) ;; addonname--1-2-3.zip
           output-path (join (fs/absolute install-dir) output-fname)] ;; /path/to/installed/addons/addonname--1.2.3.zip
       (binding [http/*cache* (cache)]
@@ -538,11 +568,12 @@
 
 (defn-spec install-addon (s/or :ok (s/coll-of ::sp/extant-file), :passed-tests true?, :error nil?)
   "downloads an addon and installs it, bypassing checks. see `install-addon-guard`."
-  [addon :addon/installable, install-dir ::sp/extant-dir, downloaded-file ::sp/extant-archive-file]
-  (try
-    (addon/install-addon addon install-dir downloaded-file)
-    (finally
-      (addon/post-install addon install-dir (get-state :cfg :preferences :addon-zips-to-keep)))))
+  [addon :addon/installable, install-dir ::sp/extant-dir, downloaded-file (s/nilable ::sp/extant-archive-file)]
+  (when downloaded-file
+    (try
+      (addon/install-addon addon install-dir downloaded-file)
+      (finally
+        (addon/post-install addon install-dir (get-state :cfg :preferences :addon-zips-to-keep))))))
 
 (def install-addon-affective
   (affects-addon-wrapper install-addon))
@@ -589,21 +620,31 @@
 ;;
 
 (defn-spec get-catalogue-location (s/or :ok :catalogue/location, :not-found nil?)
+  "returns the catalogue-location map for the given `catalogue-name`.
+  returns the catalogue-location map for the currently selected catalogue by default.
+  returns `nil` if catalogue not found or no catalogue selected and no `catalogue-name` given."
   ([]
    (get-catalogue-location (get-state :cfg :selected-catalogue)))
   ([catalogue-name keyword?]
-   (->> (get-state :cfg :catalogue-location-list) (filter #(= catalogue-name (:name %))) first)))
+   (->> (get-state :cfg :catalogue-location-list)
+        (filter #(-> % :name (= catalogue-name)))
+        first)))
+
+(defn-spec default-catalogue (s/or :ok :catalogue/location, :not-found nil?)
+  "the 'default' catalogue is the first catalogue in the list of available catalogues.
+  using the original set of catalogues that come with strongbox, this is the 'short' catalogue.
+  user can specify their own (or no) catalogues however."
+  []
+  (-> (get-state :cfg :catalogue-location-list)
+      first
+      :name ;; `:short`, typically
+      get-catalogue-location))
 
 (defn-spec current-catalogue (s/or :ok :catalogue/location, :no-catalogues nil?)
   "returns the currently selected catalogue or the first catalogue it can find.
-  returns `nil` if no catalogues available to choose from."
+  returns `nil` if no catalogues selected or none available to choose from."
   []
-  (if-let* [;; there may be nothing selected
-            catalogue (get-catalogue-location (get-state :cfg :selected-catalogue))
-            ;; there may be no default catalogue available
-            default-catalogue (get-catalogue-location (-> (get-state :cfg :catalogue-location-list) first :name))]
-           (or catalogue default-catalogue)
-           nil))
+  (or (get-catalogue-location) (default-catalogue)))
 
 (defn-spec set-catalogue-location! nil?
   [catalogue-name keyword?]
@@ -638,9 +679,29 @@
   "downloads the currently selected (or default) catalogue. 
   issues a warning if no catalogue can be downloaded"
   []
-  (if-let [catalogue (current-catalogue)]
-    (download-catalogue catalogue)
+  (if-let [catalogue-location (current-catalogue)]
+    (download-catalogue catalogue-location)
     (warn "failed to find a downloadable catalogue")))
+
+(defn-spec emergency-catalogue (s/or :static-catalogue :catalogue/catalogue, :unknown-catalogue nil?)
+  "derives the requested catalogue from the static catalogue."
+  [catalogue-location :catalogue/location]
+  (let [opts {}
+        catalogue (catalogue/read-catalogue (.getBytes (decompress-bytes static-catalogue)) opts)
+        catalogue (assoc catalogue :emergency? true)]
+
+    (warn (str "backup catalogue generated: " (:datestamp catalogue)))
+    (warn (format "remote catalogue unreachable or corrupt: %s" (:source catalogue-location)))
+
+    (case (:name catalogue-location)
+      :full catalogue
+      :short (catalogue/shorten-catalogue catalogue)
+      :curseforge (catalogue/filter-catalogue catalogue "curseforge")
+      :wowinterface (catalogue/filter-catalogue catalogue "wowinterface")
+      :tukui (catalogue/filter-catalogue catalogue "tukui")
+      nil)))
+
+;; --
 
 (defn-spec moosh-addons :addon/toc+summary+match
   "merges the data from an installed addon with it's match in the catalogue"
@@ -738,13 +799,18 @@
              catalogue-path
              {:bad-data? (fn []
                            (error "please report this! https://github.com/ogri-la/strongbox/issues")
-                           (error "catalogue *still* corrupted and cannot be loaded. try another catalogue from the 'catalogue' menu"))}))
+                           (error "remote catalogue failed to load and might be corrupt."))}))
 
-          catalogue-data (p :p2/db:catalogue:read-catalogue (catalogue/read-catalogue catalogue-path {:bad-data? bad-json-file-handler}))
-          user-catalogue-data (p :p2/db:catalogue:read-user-catalogue (catalogue/read-catalogue (paths :user-catalogue-file) {:bad-data? nil}))
+          catalogue-data (or (catalogue/read-catalogue catalogue-path {:bad-data? bad-json-file-handler})
+                             (when-not testing?
+                               (emergency-catalogue catalogue-location)))
+
+          user-catalogue-data (p :p2/db:catalogue:read-user-catalogue
+                                 (catalogue/read-catalogue (paths :user-catalogue-file) {:bad-data? nil}))
           ;; 2021-06-30: merge order changed. catalogue data is now merged over the top of the user-catalogue.
           ;; this is because the user-catalogue may now contain addons from all hosts and is likely to be out of date.
-          final-catalogue (p :p2/db:catalogue:merge-catalogues (catalogue/merge-catalogues user-catalogue-data catalogue-data))]
+          final-catalogue (p :p2/db:catalogue:merge-catalogues
+                             (catalogue/merge-catalogues user-catalogue-data catalogue-data))]
       (-> final-catalogue :addon-summary-list count (str " addons in final catalogue") info)
       final-catalogue)))
 
@@ -754,7 +820,8 @@
   []
   (if (and (not (db-catalogue-loaded?))
            (current-catalogue))
-    (let [final-catalogue (p :p2/db:catalogue (load-current-catalogue))]
+    (let [final-catalogue (p :p2/db:catalogue
+                             (load-current-catalogue))]
       (when-not (empty? final-catalogue)
         (p :p2/db:load
            (swap! state assoc :db
@@ -857,7 +924,7 @@
           has-update? (addon/updateable? addon)]
       (joblib/tick-delay 0.5)
       (when has-update?
-        (info (format "update available \"%s\"" (:version addon)))
+        (info (format "update \"%s\" available from %s" (:version addon) (:source addon)))
         (when-not (= (get-game-track) (:game-track addon))
           (warn (format "update is for '%s' and the addon directory is set to '%s'"
                         (-> addon :game-track sp/game-track-labels-map)
@@ -1158,33 +1225,30 @@
 
 (defn-spec refresh nil?
   []
-  (profile
-   {:when (get-state :profile?)}
-
-   (report "refresh")
+  (report "refresh")
 
    ;; parse toc files in install-dir. do this first so we see *something* while catalogue downloads (next)
-   (load-installed-addons)
+  (load-installed-addons)
 
    ;; downloads the big long list of addon information stored on github
-   (download-current-catalogue)
+  (download-current-catalogue)
 
    ;; load the contents of the catalogue into the database
-   (p :p2/db (db-load-catalogue))
+  (p :p2/db (db-load-catalogue))
 
    ;; match installed addons to those in catalogue
-   (match-installed-addons-with-catalogue)
+  (match-installed-addons-with-catalogue)
 
    ;; for those addons that have matches, download their details
-   (check-for-updates)
+  (check-for-updates)
 
    ;; 2019-06-30, travis is failing with 403: Forbidden. Moved to gui init
    ;;(latest-strongbox-release) ;; check for updates after everything else is done 
 
    ;; seems like a good place to preserve the etag-db
-   (save-settings)
+  (save-settings)
 
-   nil))
+  nil)
 
 ;; todo: move to ui.cli
 (defn-spec remove-addon nil?